requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
apache-airflow==2.10.2
apache-airflow-providers-common-compat==1.2.1
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.18.0
apache-airflow-providers-fab==1.4.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.1
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.7.0
appdirs==1.4.4
appnope==0.1.4
argcomplete==3.5.1
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.47
botocore==1.35.47
build==1.2.2.post1
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==3.0.4
cryptography==43.0.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.9
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
environs==9.5.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.0
filelock==3.16.1
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.6
langchain-core==0.3.14
langchain-text-splitters==0.3.1
langsmith==0.1.138
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
llama-cloud==0.1.4
llama-index==0.11.20
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.20
llama-index-embeddings-nvidia==0.2.4
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-nvidia==0.2.6
llama-index-llms-openai==0.2.16
llama-index-llms-openai-like==0.2.0
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-milvus==0.2.7
llama-parse==0.5.12
lockfile==0.12.2
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
milvus-lite==2.4.10
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.8.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
nltk==3.9.1
numpy==1.26.4
nvidia-pyindex==1.0.9
openai==1.50.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
orjson==3.10.10
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
peewee==3.17.7
pendulum==3.0.0
pexpect==4.9.0
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
prov==2.0.1
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.2
Pygments==2.18.0
PyJWT==2.9.0
pymilvus==2.4.9
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==4.3.1
PyPDF2==3.0.1
pyproject_hooks==1.2.0
PySocks==1.7.1
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scipy==1.14.1
selenium==4.25.0
setproctitle==1.3.3
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.6
streamlit==1.38.0
streamlit-chat==0.1.1
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
text-unidecode==1.3
tiktoken==0.8.0
time-machine==2.16.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.3.2
transformers==4.46.1
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
ujson==5.10.0
unicodecsv==0.14.1
universal_pathlib==0.2.5
urllib3==2.2.3
uvicorn==0.31.0
watchdog
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
yarl==1.16.0
zipp==3.20.2
